[Discussion] Fixed an issue with Transcribe Streaming's startStreamingTranscription API that could cause signature mismatch exceptions.
Because of https://github.com/netty/netty/issues/7783, writes performed by the event-loop thread can 'skip ahead' of writes
performed off of the event-loop thread. This can cause events to be sent to Transcribe out-of-order, if some are performed
on the event loop. This manifests as a signature mismatch exception because the signature of the last message is included in
the signature.

This change introduces a ChannelHandlerContext implementation that makes sure we're off of the event loop before performing a
write. If we're on the event loop, it schedules a task for the write to be performed instead of performing the task immediately.

This change isn't finalized yet, and still requires additional testing.